refactor($core): extract dataMixin, try to reuse it at build time
